pyproject.toml
[project]
name = "slackpoint"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "cachetools>=6.2.1",
    "python-dotenv>=1.2.1",
    "slack-bolt>=1.26.0",
]
